Change indexing from int to Decimal
This commit changes the indexing used in make_cooptimisation_model from integers to a Decimal type.

The Decimal type is used to avoid floating point errors. Specifically, the float 0.4 does not equal the resulting float of 0.6-0.2. This caused problems whilst testing as this meant that soc[0.6-0.2] would not access soc[0.4] as intended, causing an error.