Change data processing pipeline
The previous method of processing the FCAS price data first downloaded the daily dispatch reports via a script, which was then processed on import of `src.data`. A DataFrame of a single day of SA FCAS prices could then be accessed by a module attribute `sa_price_df`. However, this was inconvenient for when data from other days were needed, as the procedure for data extraction was hard-coded.

This commit fixes this by replacing the code of data.py to have only three functions for the data processing and retrieval of SA FCAS prices:

- download_daily_reports (code from download.py)
- extract_sa_fcas_prices (code from old data.py initialisation)
- get_sa_fcas_prices

download.py has also been edited accordingly to output a csv of SA FCAS prices under data/.