Ensure FCAS prices are sorted upon extraction
This commit ensures that FCAS prices are sorted in the DataFrame returned by extract_sa_fcas_prices(). As each of the public dispatch CSVs were found to be sorted, this meant iterating through each CSVs in order of filename when compiling all SA FCAS prices. To ensure this, an `assert` statement has also been added to check whether the data is sorted.