Refactor make_cooptimisation_model
Previously, each variable for each service in make_cooptimisation_model was defined manually. This created lengthy code for the objectives and constraints, for example 6 decision variables had to be manually added in the objective and 6 separate calls to `m.addConstrs` had to be made for each of the big M constraints. This commit thus makes several changes to refactor the model for readability:

- the 6 contingency FCAS types are defined as indices
- the decision variables are now defined based on a pair of FCAS type and trading interval
- prices are now provided into the model via a `prices` dictionary instead of 6 separate price parameters

Since variable names are changed, the utility functions `tabulate_solution` and `show_solution` have also been updated.

To ensure that experiments run correctly, the `delta_t` value for `lower_6_sec` has been set to `5 / 60` instead of the correct value of `6 / 60 / 60`, which was a mistake from in the previous version of make_cooptimisation_model.