Include price scenarios in make_scenario_model
Add price scenarios to make_scenario_model. Price scenarios are combined with enablement scenarios (via pointwise multiplication) to create a scenario. This commit adds the parameter `scenario_combine_method` to indicate whether to combine all combinations of price and enablement (with 'product') or to combine corresponding price and enablement scenarios in order of their appearance (with 'zip').