Remove the num_scenarios parameter
Since the number of price and enablement scenarios can be inferred from the passed values to parameters `prices` and `enablement_scenarios`, and since the number of combined scenarios will depend on `scenario_combine_method`, the `num_scenarios` parameter has been removed. This also allows an unequal number of price and enablement scenarios to be passed to the function (which was impossible in previous versions since both sets of scenarios needed to contain `num_scenarios` scenarios).