Optimise scenario probability calculation
This commit makes an optimisation on the scenario probability calculation. Instead of re-calculating the probability upon flipping the bit of a scenario, the old probability can be divided by the probability of success of the bit, then multiplied by the inverse of that probability. This gives the probability of the new scenario in constant time.