Change to log probability, fix flip_scenario
This commit makes two changes. First, instead of sorting scenarios in terms of probabilities, the log of the probabilities is used instead. This prevents precision issues that may arise with larger scenario lengths due to multiplying many fractional numbers (e.g. such as length 288 scenarios). Second, flip_scenario() has been changed to fix an issue on indexing: the integer scenarios have to be indexed MSB first instead of LSB first as in the previous revision. This ensures that the indexed bits correspond to the correct scenario probability.