Pre-calculate decision variable constants
The objective function of the stochastic model can be expressed as the sum of each decision variable multiplied by a corresponding constant. This commit contains changes to the construction of the objective make_scenario_model to calculate this constant outside of setObjective. From initial testing, this change provides a speedup for the model construction routine, and solutions appear to be the same. However, further investigation into precision issues may have to be done.