Move constants calculation to external function
This commit moves the calculation of the objective function constants, called `scenario_consts`, to an external function src.utils.calc_scenario_consts.

This commit also simplifies the calculation when scenario_combine_method is 'product'. This calculation is equivalent to the previous calculation due to the distributivity rule. It saves memory by avoiding the storage of all scenarios produced and saves time by avoiding calculation of the weighted sum of all possible scenarios.